+++ Samples/kotlin/src/main/java/com/phunware/kotlin/sample/LocationSharingActivity.kt
@@ -1,5 +1,31 @@
 package com.phunware.kotlin.sample
 
+/* Copyright (C) 2018 Phunware, Inc.
+
+Permission is hereby granted, free of charge, to any person obtaining
+a copy of this software and associated documentation files (the
+"Software"), to deal in the Software without restriction, including
+without limitation the rights to use, copy, modify, merge, publish,
+distribute, sublicense, and/or sell copies of the Software, and to
+permit persons to whom the Software is furnished to do so, subject to
+the following conditions:
+
+The above copyright notice and this permission notice shall be
+included in all copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+NONINFRINGEMENT. IN NO EVENT SHALL Phunware, Inc. BE LIABLE FOR ANY
+CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
+TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
+SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+
+Except as contained in this notice, the name of Phunware, Inc. shall
+not be used in advertising or otherwise to promote the sale, use or
+other dealings in this Software without prior written authorization
+from Phunware, Inc. */
+
 import android.content.Context
 import android.graphics.Bitmap
 import android.graphics.Canvas
@@ -152,6 +178,9 @@ class LocationSharingActivity : AppCompatActivity(), OnPhunwareMapReadyCallback,
                         spinnerAdapter.clear()
                         spinnerAdapter.addAll(building.buildingOptions.floors)
 
+                        // Add a listener to monitor floor switches
+                        mapManager.addFloorChangedListener(this@LocationSharingActivity)
+
                         // Initialize a location provider
                         setManagedLocationProvider(building)
 
@@ -167,7 +196,6 @@ class LocationSharingActivity : AppCompatActivity(), OnPhunwareMapReadyCallback,
 
                         // Start sharing location with other users
                         startLocationSharing()
-                        mapManager.startRetrievingSharedLocations(this@LocationSharingActivity)
                     }
 
                     override fun onFailure(throwable: Throwable) {
@@ -237,9 +265,7 @@ class LocationSharingActivity : AppCompatActivity(), OnPhunwareMapReadyCallback,
         // Remove ourself from this list
         val sharedLocationListWithoutSelf = ArrayList<SharedLocation>()
         for (location in sharedLocationList) {
-            val deviceId = PwCoreSession.getInstance()
-                    .sessionData.deviceId
-
+            val deviceId = PwCoreSession.getInstance().sessionData.deviceId
             if (deviceId != location.deviceId) {
                 sharedLocationListWithoutSelf.add(location)
             }
@@ -266,8 +292,8 @@ class LocationSharingActivity : AppCompatActivity(), OnPhunwareMapReadyCallback,
                 removePersonDot(staleDevice)
             }
 
-            for (p in sharedLocationListWithoutSelf) {
-                updatePersonDot(p)
+            for (userLocation in sharedLocationListWithoutSelf) {
+                updatePersonDot(userLocation)
             }
         }
     }